PlayWrightCicada8/requirements.txt
# Зависимости проекта
playwright==1.44.0      # Библиотека для автоматизации браузера
pytest==8.2.2          # Фреймворк для написания и запуска тестов
python-dotenv==1.0.1   # Для чтения переменных окружения из .env
pytest-xdist==3.6.1    # Для параллельного запуска тестов